+++ pkg/mimir/grpc_push_check_test.go
@@ -34,18 +34,17 @@ func TestGrpcInflightMethodLimiter(t *testing.T) {
 
 		require.Panics(t, func() {
 			// In practice, this will not be called, since l.RPCCallStarting() for ingester push returns error if there's no ingester.
-			l.RPCCallFinished(context.WithValue(ctx, pushTypeCtxKey, ingesterPush))
+			l.RPCCallFinished(context.WithValue(ctx, pushTypeCtxKey, pushTypeIngester))
 		})
 
 		require.Panics(t, func() {
 			// In practice, this will not be called, since l.RPCCallStarting() distributor push returns error if there's no distributor.
-			l.RPCCallFinished(context.WithValue(ctx, pushTypeCtxKey, distributorPush))
+			l.RPCCallFinished(context.WithValue(ctx, pushTypeCtxKey, pushTypeDistributor))
 		})
 	})
 
-	t.Run("ingester push receiver", func(t *testing.T) {
+	t.Run("ingester push receiver, wrong method name", func(t *testing.T) {
 		m := &mockIngesterReceiver{}
-
 		l := newGrpcInflightMethodLimiter(func() ingesterPushReceiver { return m }, nil)
 
 		ctx, err := l.RPCCallStarting(context.Background(), "test", nil)
@@ -55,24 +54,82 @@ func TestGrpcInflightMethodLimiter(t *testing.T) {
 		})
 		require.Equal(t, 0, m.startCalls)
 		require.Equal(t, 0, m.finishCalls)
+	})
 
-		ctx, err = l.RPCCallStarting(context.Background(), ingesterPushMethod, nil)
+	t.Run("ingester push receiver, check returns error", func(t *testing.T) {
+		m := &mockIngesterReceiver{}
+		l := newGrpcInflightMethodLimiter(func() ingesterPushReceiver { return m }, nil)
+
+		m.returnError = errors.New("hello there")
+		ctx, err := l.RPCCallStarting(context.Background(), ingesterPushMethod, nil)
+		require.Error(t, err)
+		_, ok := status.FromError(err)
+		require.True(t, ok)
+		require.Nil(t, ctx.Value(pushTypeCtxKey)) // Original context expected in case of errors.
+	})
+
+	t.Run("ingester push receiver, without size", func(t *testing.T) {
+		m := &mockIngesterReceiver{}
+		l := newGrpcInflightMethodLimiter(func() ingesterPushReceiver { return m }, nil)
+
+		ctx, err := l.RPCCallStarting(context.Background(), ingesterPushMethod, nil)
 		require.NoError(t, err)
 		require.Equal(t, 1, m.startCalls)
+		require.Equal(t, int64(0), m.startBytes)
 		require.Equal(t, 0, m.finishCalls)
+		require.Equal(t, int64(0), m.finishBytes)
 
 		require.NotPanics(t, func() {
 			l.RPCCallFinished(ctx)
 		})
 		require.Equal(t, 1, m.startCalls)
+		require.Equal(t, int64(0), m.startBytes)
 		require.Equal(t, 1, m.finishCalls)
+		require.Equal(t, int64(0), m.finishBytes)
+	})
 
-		m.returnError = errors.New("hello there")
-		ctx, err = l.RPCCallStarting(context.Background(), ingesterPushMethod, nil)
-		require.Error(t, err)
-		_, ok := status.FromError(err)
-		require.True(t, ok)
-		require.Nil(t, ctx.Value(pushTypeCtxKey)) // Original context expected in case of errors.
+	t.Run("ingester push receiver, with size provided", func(t *testing.T) {
+		m := &mockIngesterReceiver{}
+		l := newGrpcInflightMethodLimiter(func() ingesterPushReceiver { return m }, nil)
+
+		ctx, err := l.RPCCallStarting(context.Background(), ingesterPushMethod, metadata.New(map[string]string{
+			grpcutil.MetadataMessageSize: "123456",
+		}))
+		require.NoError(t, err)
+		require.Equal(t, 1, m.startCalls)
+		require.Equal(t, int64(123456), m.startBytes)
+		require.Equal(t, 0, m.finishCalls)
+		require.Equal(t, int64(0), m.finishBytes)
+
+		require.NotPanics(t, func() {
+			l.RPCCallFinished(ctx)
+		})
+		require.Equal(t, 1, m.startCalls)
+		require.Equal(t, int64(123456), m.startBytes)
+		require.Equal(t, 1, m.finishCalls)
+		require.Equal(t, int64(123456), m.finishBytes)
+	})
+
+	t.Run("ingester push receiver, with wrong size", func(t *testing.T) {
+		m := &mockIngesterReceiver{}
+		l := newGrpcInflightMethodLimiter(func() ingesterPushReceiver { return m }, nil)
+
+		ctx, err := l.RPCCallStarting(context.Background(), ingesterPushMethod, metadata.New(map[string]string{
+			grpcutil.MetadataMessageSize: "wrong",
+		}))
+		require.NoError(t, err)
+		require.Equal(t, 1, m.startCalls)
+		require.Equal(t, int64(0), m.startBytes)
+		require.Equal(t, 0, m.finishCalls)
+		require.Equal(t, int64(0), m.finishBytes)
+
+		require.NotPanics(t, func() {
+			l.RPCCallFinished(ctx)
+		})
+		require.Equal(t, 1, m.startCalls)
+		require.Equal(t, int64(0), m.startBytes)
+		require.Equal(t, 1, m.finishCalls)
+		require.Equal(t, int64(0), m.finishBytes)
 	})
 
 	t.Run("distributor push via httpgrpc", func(t *testing.T) {
@@ -167,17 +224,21 @@ func TestGrpcInflightMethodLimiter(t *testing.T) {
 
 type mockIngesterReceiver struct {
 	startCalls  int
+	startBytes  int64
 	finishCalls int
+	finishBytes int64
 	returnError error
 }
 
-func (i *mockIngesterReceiver) StartPushRequest() error {
+func (i *mockIngesterReceiver) StartPushRequest(size int64) error {
 	i.startCalls++
+	i.startBytes += size
 	return i.returnError
 }
 
-func (i *mockIngesterReceiver) FinishPushRequest() {
+func (i *mockIngesterReceiver) FinishPushRequest(size int64) {
 	i.finishCalls++
+	i.finishBytes += size
 }
 
 type mockDistributorReceiver struct {